Added tests for getQuestion utility function, refactored non-selectors into utils.js
