+++ frontend/test/unit/reference/utils.spec.js
@@ -2,7 +2,10 @@ import {
     tryFetchData,
     tryUpdateData,
     tryUpdateFields,
-    separateTablesBySchema
+    buildBreadcrumbs,
+    databaseToForeignKeys,
+    separateTablesBySchema,
+    getQuestion
 } from 'metabase/reference/utils';
 
 describe("Reference utils.js", () => {
@@ -139,6 +142,116 @@ describe("Reference utils.js", () => {
         });
     });
 
+    describe("databaseToForeignKeys()", () => {
+        it("should build foreignKey viewmodels from database", () => {
+            const database = {
+                tables_lookup: {
+                    1: {
+                        id: 1,
+                        display_name: 'foo',
+                        schema: 'PUBLIC',
+                        fields_lookup: {
+                            1: {
+                                id: 1,
+                                special_type: 'id',
+                                display_name: 'bar',
+                                description: 'foobar'
+                            }
+                        }
+                    },
+                    2: {
+                        id: 2,
+                        display_name: 'bar',
+                        schema: 'public',
+                        fields_lookup: {
+                            2: {
+                                id: 2,
+                                special_type: 'id',
+                                display_name: 'foo',
+                                description: 'barfoo'
+                            }
+                        }
+                    },
+                    3: {
+                        id: 3,
+                        display_name: 'boo',
+                        schema: 'TEST',
+                        fields_lookup: {
+                            3: {
+                                id: 3,
+                                display_name: 'boo',
+                                description: 'booboo'
+                            }
+                        }
+                    }
+                }
+            };
+
+            const foreignKeys = databaseToForeignKeys(database);
+
+            expect(foreignKeys).toEqual({
+                1: { id: 1, name: 'Public.foo → bar', description: 'foobar' },
+                2: { id: 2, name: 'bar → foo', description: 'barfoo' }
+            });
+        });
+    });
+
+    describe("buildBreadcrumbs()", () => {
+        const section1 = {
+            id: 1,
+            breadcrumb: 'section1'
+        };
+
+        const section2 = {
+            id: 2,
+            breadcrumb: 'section2',
+            parent: section1
+        };
+
+        const section3 = {
+            id: 3,
+            breadcrumb: 'section3',
+            parent: section2
+        };
+
+        const section4 = {
+            id: 4,
+            breadcrumb: 'section4',
+            parent: section3
+        };
+
+        const section5 = {
+            id: 5,
+            breadcrumb: 'section5',
+            parent: section4
+        };
+
+        it("should build correct breadcrumbs from parent section", () => {
+            const breadcrumbs = buildBreadcrumbs(section1);
+            expect(breadcrumbs).toEqual([
+                [ 'section1' ]
+            ]);
+        });
+
+        it("should build correct breadcrumbs from child section", () => {
+            const breadcrumbs = buildBreadcrumbs(section3);
+            expect(breadcrumbs).toEqual([
+                [ 'section1', 1 ],
+                [ 'section2', 2 ],
+                [ 'section3' ]
+            ]);
+        });
+
+        it("should keep at most 3 highest level breadcrumbs", () => {
+            const breadcrumbs = buildBreadcrumbs(section5);
+            expect(breadcrumbs).toEqual([
+                [ 'section3', 3 ],
+                [ 'section4', 4 ],
+                [ 'section5' ]
+            ]);
+        });
+    });
+
     describe("tablesToSchemaSeparatedTables()", () => {
         it("should add schema separator to appropriate locations", () => {
             const tables = {
@@ -170,6 +283,164 @@ describe("Reference utils.js", () => {
                 ["foo", { id: 1, name: 'table1', schema: 'foo' }],
                 { id: 5, name: 'table5', schema: 'foo' },
             ]);
+        });
+    });
+
+    describe("getQuestion()", () => {
+        const getNewQuestion = ({
+            database = 1,
+            table = 2,
+            display = "table",
+            aggregation = [ "rows" ],
+            breakout = [],
+            filter = []
+        }) => ({
+            "name": null,
+            "public_perms": 0,
+            "display": display,
+            "visualization_settings": {},
+            "dataset_query": {
+                "database": database,
+                "type": "query",
+                "query": {
+                    "source_table": table,
+                    "aggregation": aggregation,
+                    "breakout": breakout,
+                    "filter": filter
+                }
+            }
+        });
+
+        it("should generate correct question for table raw data", () => {
+            const question = getQuestion({
+                dbId: 3,
+                tableId: 4
+            });
+
+            expect(question).toEqual(getNewQuestion({
+                database: 3,
+                table: 4
+            }));
+        });
+
+        it("should generate correct question for table counts", () => {
+            const question = getQuestion({
+                dbId: 3,
+                tableId: 4,
+                getCount: true
+            });
+
+            expect(question).toEqual(getNewQuestion({
+                database: 3,
+                table: 4,
+                aggregation: [ "count" ]
+            }));
+        });
+
+        it("should generate correct question for field raw data", () => {
+            const question = getQuestion({
+                dbId: 3,
+                tableId: 4,
+                fieldId: 5
+            });
+
+            expect(question).toEqual(getNewQuestion({
+                database: 3,
+                table: 4,
+                breakout: [ 5 ]
+            }));
+        });
+
+        it("should generate correct question for field group by bar chart", () => {
+            const question = getQuestion({
+                dbId: 3,
+                tableId: 4,
+                fieldId: 5,
+                getCount: true,
+                visualization: 'bar'
+            });
+
+            expect(question).toEqual(getNewQuestion({
+                database: 3,
+                table: 4,
+                display: 'bar',
+                breakout: [ 5 ],
+                aggregation: [ "count" ]
+            }));
+        });
+
+        it("should generate correct question for field group by pie chart", () => {
+            const question = getQuestion({
+                dbId: 3,
+                tableId: 4,
+                fieldId: 5,
+                getCount: true,
+                visualization: 'pie'
+            });
+
+            expect(question).toEqual(getNewQuestion({
+                database: 3,
+                table: 4,
+                display: 'pie',
+                breakout: [ 5 ],
+                aggregation: [ "count" ]
+            }));
+        });
+
+        it("should generate correct question for metric raw data", () => {
+            const question = getQuestion({
+                dbId: 1,
+                tableId: 2,
+                metricId: 3
+            });
+
+            expect(question).toEqual(getNewQuestion({
+                aggregation: [ "METRIC", 3 ]
+            }));
+        });
+
+        it("should generate correct question for metric group by fields", () => {
+            const question = getQuestion({
+                dbId: 1,
+                tableId: 2,
+                fieldId: 4,
+                metricId: 3
+            });
+
+            expect(question).toEqual(getNewQuestion({
+                aggregation: [ "METRIC", 3 ],
+                breakout: [ 4 ]
+            }));
+        });
+
+        it("should generate correct question for segment raw data", () => {
+            const question = getQuestion({
+                dbId: 2,
+                tableId: 3,
+                segmentId: 4
+            });
+
+            expect(question).toEqual(getNewQuestion({
+                database: 2,
+                table: 3,
+                filter: [ "AND", [ "SEGMENT", 4 ] ]
+            }));
+        });
+
+        it("should generate correct question for segment counts", () => {
+            const question = getQuestion({
+                dbId: 2,
+                tableId: 3,
+                segmentId: 4,
+                getCount: true
+            });
+
+            expect(question).toEqual(getNewQuestion({
+                database: 2,
+                table: 3,
+                aggregation: [ "count" ],
+                filter: [ "AND", [ "SEGMENT", 4 ] ]
+            }));
         });
     });
 });
